+++ packages/bruno-common/src/utils/index.ts
@@ -42,3 +42,9 @@ export {
 export {
   toDisplayString
 } from './string';
+
+export {
+  toEntries,
+  buildHeaderSections,
+  flattenHeaderSections
+} from './headers-source';
